requirements.txt
python-telegram-bot==21.10
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
aiohttp==3.11.11
pyngrok==7.2.2
